refactor(assistant): extract ThreadWorkspace, inline conflict confirm, restyle inbox rows
